IOTEX-136 Create multichain dir and split mainchain and subchain protocols
